transpile: Move `SomeId` to main `c_ast` module
